Shared tasks & add/remove to list.

+++ views.py
@@ -19,7 +19,12 @@ def tasks_page():
     db = current_app.config["db"]
     if request.method == "GET":
         tasks = db.get_tasks(current_user.username)
-        return render_template("tasks.html", tasks=sorted(tasks))
+        shared_tasks = []
+        for task in tasks[:]:
+            if db.count_task_user_relation(task[0])[0] >= 2:
+                tasks.remove(task)
+                shared_tasks.append(task)
+        return render_template("tasks.html", tasks=sorted(tasks), shared_tasks=sorted(shared_tasks))
     else:
         form_task_keys = request.form.getlist("task_keys")
         for form_task_key in form_task_keys:
@@ -37,7 +42,12 @@ def task_page(task_key):
     task = db.get_task(task_key, current_user.username)
     if task is None:
         abort(404)
-    return render_template("task.html", task=task)
+    task_share = db.get_task_share(task_key)
+    if len(task_share) <= 1:
+        task_share = None
+    else:
+        task_share = sorted(task_share)
+    return render_template("task.html", task=task, task_share=task_share)
 
 
 @login_required
@@ -50,6 +60,9 @@ def task_add_page():
         db = current_app.config["db"]
         task_key = db.add_task(task)
         db.add_task_user_relation(task_key, current_user.username)
+        share = [x for x in form.data["share"].split(",") if x != "" and x != " "]
+        for username in share:
+            db.add_task_user_relation(task_key, username)
         flash("Task added.")
         return redirect(url_for("task_page", task_key=task_key))
     return render_template("task_edit.html", form=form)
@@ -63,15 +76,31 @@ def task_edit_page(task_key):
         abort(404)
     form = TaskEditForm()
     if form.validate_on_submit():
-        name = form.data["name"]
-        description = form.data["description"]
-        task = Task(name, description=description)
-        db.update_task(task_key, task)
+        # add status location etc.
+        # make for list?
+        share = [x for x in form.data["share"].split(",") if x != "" and x != " "]
+        db.delete_task_user_relation(task_key)
+        for username in share:
+            db.add_task_user_relation(task_key, username)
+        if not db.is_exist_task_user_relation(task_key):
+            db.delete_task(task_key)
+        else:
+            name = form.data["name"]
+            description = form.data["description"]
+            task = Task(name, description=description)
+            db.update_task(task_key, task)
         flash("Task data updated.")
+        if current_user.username not in share:
+            return redirect(url_for("tasks_page"))
         return redirect(url_for("task_page", task_key=task_key))
     form.name.data = task.name
     form.description.data = task.description if task.description else ""
-    return render_template("task_edit.html", form=form)
+    task_share = db.get_task_share(task_key)
+    if len(task_share) < 1:
+        task_share = None
+    else:
+        task_share = sorted(task_share)
+    return render_template("task_edit.html", form=form, task_share=task_share)
 
 
 @login_required
@@ -99,15 +128,28 @@ def list_page(list_key):
         if list is None:
             abort(404)
         tasks = db.get_tasks_with_list(list_key)
-        return render_template("list.html", list=list, tasks=sorted(tasks))
+        shared_tasks = []
+        for task in tasks[:]:
+            if db.count_task_user_relation(task[0])[0] >= 2:
+                tasks.remove(task)
+                shared_tasks.append(task)
+        return render_template("list.html", list=list, tasks=sorted(tasks), shared_tasks=sorted(shared_tasks))
     else:
         form_task_keys = request.form.getlist("task_keys")
-        for form_task_key in form_task_keys:
-            task = db.get_task(form_task_key, current_user.username)
-            if task is None:
-                abort(404)
-            db.delete_task(int(form_task_key))
-        flash("%(num)d tasks deleted." % {"num": len(form_task_keys)})
+        if "remove" in request.form:
+            for form_task_key in form_task_keys:
+                task = db.get_task(form_task_key, current_user.username)
+                if task is None:
+                    abort(404)
+                db.task_remove_list(int(form_task_key))
+            flash("%(num)d tasks removed from this list." % {"num": len(form_task_keys)})
+        elif "delete" in request.form:
+            for form_task_key in form_task_keys:
+                task = db.get_task(form_task_key, current_user.username)
+                if task is None:
+                    abort(404)
+                db.delete_task(int(form_task_key))
+            flash("%(num)d tasks deleted." % {"num": len(form_task_keys)})
         return redirect(url_for("list_page", list_key=list_key))
 
 
@@ -146,7 +188,7 @@ def list_edit_page(list_key):
 
 
 @login_required
-def list_add_task_page(list_key):
+def list_new_task_page(list_key):
     form = TaskEditForm()
     if form.validate_on_submit():
         name = form.data["name"]
@@ -155,9 +197,31 @@ def list_add_task_page(list_key):
         db = current_app.config["db"]
         task_key = db.add_task_with_list(task)
         db.add_task_user_relation(task_key, current_user.username)
+        share = [x for x in form.data["share"].split(",") if x != "" and x != " "]
+        for username in share:
+            db.add_task_user_relation(task_key, username)
         flash("Task added.")
         return redirect(url_for("list_page", list_key=list_key))
     return render_template("task_edit.html", form=form)
+
+
+@login_required
+def list_add_task_page(list_key):
+    db = current_app.config["db"]
+    if request.method == "GET":
+        tasks = db.get_tasks(current_user.username)
+        shared_tasks = []
+        for task in tasks[:]:
+            if db.count_task_user_relation(task[0])[0] >= 2:
+                tasks.remove(task)
+                shared_tasks.append(task)
+        return render_template("add_task_to_list.html", tasks=sorted(tasks), shared_tasks=sorted(shared_tasks))
+    else:
+        form_task_keys = request.form.getlist("task_keys")
+        for form_task_key in form_task_keys:
+            db.task_add_list(list_key, form_task_key)
+        flash("%(num)d tasks added to this list." % {"num": len(form_task_keys)})
+        return redirect(url_for("list_page", list_key=list_key))
 
 
 def login_page():
